test(rx_lpc): drop 10 redundant uint32_t casts on k_lpc_wake_* enum constants

+++ star-rx72n-firmware/tests/test_rx_lpc.c
@@ -96,7 +96,7 @@ static void test_init_consumes_injected_state(void)
   test_setup();
 
   rx_lpc_test_set_deep_standby_wake(true);
-  rx_lpc_test_set_pending_wake_flags((uint32_t)k_lpc_wake_rtc_alarm);
+  rx_lpc_test_set_pending_wake_flags(k_lpc_wake_rtc_alarm);
 
   rx_err_t err = rx_lpc_init();
   TEST_ASSERT_EQUAL(k_rx_ok, err);
@@ -146,7 +146,7 @@ static void test_enter_deep_standby_before_init_returns_not_initialized(void)
   test_setup();
 
   rx_err_t err =
-    rx_lpc_enter_deep_software_standby((uint32_t)k_lpc_wake_irq0, k_lpc_deep_ram_usb_on, false);
+    rx_lpc_enter_deep_software_standby(k_lpc_wake_irq0, k_lpc_deep_ram_usb_on, false);
   TEST_ASSERT_EQUAL(k_rx_err_not_initialized, err);
 }
 
@@ -275,7 +275,7 @@ static void test_enter_deep_standby_unknown_bits_returns_invalid_arg(void)
   TEST_ASSERT_EQUAL(k_rx_ok, rx_lpc_init());
 
   /* Bit 31 is not assigned in DPSIER3 (only bit 24 = CAN1 RX) */
-  const uint32_t bad_mask = (uint32_t)k_lpc_wake_irq0 | 0x80000000U;
+  const uint32_t bad_mask = k_lpc_wake_irq0 | 0x80000000U;
   rx_err_t       err = rx_lpc_enter_deep_software_standby(bad_mask, k_lpc_deep_ram_usb_on, false);
   TEST_ASSERT_EQUAL(k_rx_err_invalid_arg, err);
 }
@@ -285,7 +285,7 @@ static void test_enter_deep_standby_invalid_power_returns_invalid_arg(void)
   test_setup();
   TEST_ASSERT_EQUAL(k_rx_ok, rx_lpc_init());
 
-  rx_err_t err = rx_lpc_enter_deep_software_standby((uint32_t)k_lpc_wake_irq0,
+  rx_err_t err = rx_lpc_enter_deep_software_standby(k_lpc_wake_irq0,
                                                     (rx_lpc_deep_power_t)0xFFU,
                                                     false);
   TEST_ASSERT_EQUAL(k_rx_err_invalid_arg, err);
@@ -296,7 +296,7 @@ static void test_enter_deep_standby_success_records_last_mode(void)
   test_setup();
   TEST_ASSERT_EQUAL(k_rx_ok, rx_lpc_init());
 
-  const uint32_t mask = (uint32_t)k_lpc_wake_irq0 | (uint32_t)k_lpc_wake_rtc_alarm;
+  const uint32_t mask = k_lpc_wake_irq0 | k_lpc_wake_rtc_alarm;
   rx_err_t       err  = rx_lpc_enter_deep_software_standby(mask, k_lpc_deep_ram_usb_off, true);
   TEST_ASSERT_EQUAL(k_rx_ok, err);
   TEST_ASSERT_EQUAL((int)k_lpc_mode_deep_software_standby, (int)rx_lpc_test_get_last_mode());
@@ -308,7 +308,7 @@ static void test_enter_deep_standby_accepts_full_wake_mask(void)
   TEST_ASSERT_EQUAL(k_rx_ok, rx_lpc_init());
 
   rx_err_t err =
-    rx_lpc_enter_deep_software_standby((uint32_t)k_lpc_wake_all_mask, k_lpc_deep_lvd_off, false);
+    rx_lpc_enter_deep_software_standby(k_lpc_wake_all_mask, k_lpc_deep_lvd_off, false);
   TEST_ASSERT_EQUAL(k_rx_ok, err);
 }
 
@@ -319,13 +319,13 @@ static void test_enter_deep_standby_accepts_each_deep_power_level(void)
 
   TEST_ASSERT_EQUAL(
     k_rx_ok,
-    rx_lpc_enter_deep_software_standby((uint32_t)k_lpc_wake_irq0, k_lpc_deep_ram_usb_on, false));
+    rx_lpc_enter_deep_software_standby(k_lpc_wake_irq0, k_lpc_deep_ram_usb_on, false));
   TEST_ASSERT_EQUAL(
     k_rx_ok,
-    rx_lpc_enter_deep_software_standby((uint32_t)k_lpc_wake_irq0, k_lpc_deep_ram_usb_off, false));
+    rx_lpc_enter_deep_software_standby(k_lpc_wake_irq0, k_lpc_deep_ram_usb_off, false));
   TEST_ASSERT_EQUAL(
     k_rx_ok,
-    rx_lpc_enter_deep_software_standby((uint32_t)k_lpc_wake_irq0, k_lpc_deep_lvd_off, false));
+    rx_lpc_enter_deep_software_standby(k_lpc_wake_irq0, k_lpc_deep_lvd_off, false));
 }
 
 /* =============================================================================
@@ -377,7 +377,7 @@ static void test_get_wake_flags_returns_injected_value(void)
   test_setup();
 
   const uint32_t injected =
-    (uint32_t)k_lpc_wake_irq3 | (uint32_t)k_lpc_wake_rtc_alarm | (uint32_t)k_lpc_wake_can1_rx;
+    k_lpc_wake_irq3 | k_lpc_wake_rtc_alarm | k_lpc_wake_can1_rx;
   rx_lpc_test_set_pending_wake_flags(injected);
   TEST_ASSERT_EQUAL(k_rx_ok, rx_lpc_init());
